personal_wallet/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
packaging==23.1
pipreqs==0.4.13
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
ruff==0.0.287
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
whitenoise==6.5.0
yarg==0.1.9
